Accepts s3 options when updating lambda functions (#39)
The deploy-lambda-function now takes s3 options and will upload the assets to an s3 bucket prior to updating the code.  After the lambda is updated, the lambda will also publish a new version so as to lock the changes in.

To disable uploading to s3 but still using the s3 options, set `upload-to-s3` to `false`.
To disable publishing the lambda version, set `publish-version` to `false`.

If no s3-bucket is provided, the upload to s3 will not happen.